+++ GLG3D.lib/source/Shader.cpp
@@ -846,7 +846,7 @@ void VertexAndPixelShader::validateArgList(const ArgList& args) const {
 
     }
 
-    if (numVariables < args.argTable.size()) {
+    if (numVariables < static_cast<int>(args.argTable.size())) {
         // Some variables were unused.  Figure out which they were.
         Table<std::string, ArgList::Arg>::Iterator arg = args.argTable.begin();
         Table<std::string, ArgList::Arg>::Iterator end = args.argTable.end();
